Move seed settings to same folder
